+++ src/Faithlife.Build/MSBuildVersion.cs
@@ -1,0 +1,18 @@
+namespace Faithlife.Build
+{
+	/// <summary>
+	/// The version of MSBuild to run.
+	/// </summary>
+	public enum MSBuildVersion
+	{
+		/// <summary>
+		/// Visual Studio 2017.
+		/// </summary>
+		VS2017,
+
+		/// <summary>
+		/// Visual Studio 2019.
+		/// </summary>
+		VS2019,
+	}
+}